Add files via upload
PNX AI project is aiming to be the future brain of the data on the blockchain. Developing three core products, PNX AI will deploy AI and Machine learning to enhance the user experience on the blockchain. Its token is deflationary and taxed 10% from each transaction. 5% goes to auto-generating liquidity and 5% is distributed to reward current holders.